ml_pipeline/pyproject.toml
[project]
name = "hotel-ml-pipeline"
version = "0.1.0"
description = "ML pricing pipeline for hotel booking system"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.26.0",
    "sagemaker>=2.150.0",
    "xgboost>=1.7.0",
    "scikit-learn>=1.2.0",
    "pandas>=1.5.0",
    "numpy>=1.24.0",
    "joblib>=1.2.0",
    "psycopg2-binary>=2.9.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py39"